Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> simulator "entry points" are: 
>  - old QEMU (old ansible infra)
>  - new QEMU (new builder infra)
> and host tests:
>  - simulator/eVSSIM/tests/host/log_mgr_tests.cc
>  - simulator/eVSSIM/tests/host/sector_tests.cc
>  - simulator/eVSSIM/tests/host/object_tests.cc   
>  - simulator/eVSSIM/tests/host/ssd_io_emulator_tests.cc
> 
> ATM, INIT_LOG_MANAGER is called from the following places:
> simulator/eVSSIM/QEMU/hw/nvme_storage.c:679:	INIT_LOG_MANAGER(); 
> simulator/eVSSIM/tests/host/base_emulator_tests.h:111:                INIT_LOG_MANAGER();
> simulator/eVSSIM/tests/host/object_tests.cc:104:                INIT_LOG_MANAGER();
> qemu/hw/vssim/vssim.c:100:        INIT_LOG_MANAGER();
> 
> I guess there is no point to call INIT_LOG_MANAGER from the object_tests.cc, since one test suite is mapped to one testing domain. That leaves 3 must have places to initialize INIT_LOG_MANAGER:
> - qemu/hw/vssim/vssim.c
> - simulator/eVSSIM/QEMU/hw/nvme_storage.c
> - simulator/eVSSIM/tests/host/ssd_io_emulator_tests.cc (base_emulator_tests.h)

If we call init from: qemu/hw/vssim/vssim.c
We will not be able to use the libwebsockets because the port is already in use and thus we have the binding issue.
I am no libwebsocket expert and not familiar with it much, but they say that the work mode should be one context with many vhosts. I am not sure how it is possible since you can't create vhost without having context at your hand. We use multi-threaded solution which makes it a bit tough.
anyway, during CI run it is ok to have no change - logger will work fine it seems.
Otherwise if we initiate docker-run-qemu.sh first, it will create context and bind to 2003 port and stop others from binding to this port by returning the error we see.
Now i wonder if we should have some design changes to logger.

Patch-set: 1