rfc: Dynamic namespace support in QEMU
This is a WIP, non-compiling, skeleton implementation of how dynamic
namespaces would work in QEMU to start a conversation and debate on how
to go about it.

Currently, it is extending `BlockDriver` with new functionality to let
drivers decide on how they want to handle namespaces.

It works by having a master `BlockDriver` that manages the namespaces
and saves the state to make it persistent accross resets. To get a
particular namespace, devices call `bdrv_get_ns()` to get a child
`BlockDriver` that implements the usual functions expected of a
`BlockDriver`.

It is not complete as more functions are needed like
`bdrv_ns_get_allocated_count()` and `bdrv_ns_get_info()` and probably
more.

I am not really fond on this particular approach as the master
BlockDriver that handles namespace management isn't really a
"BlockDriver" in the usual sense, i.e. it doesn't have a use for
any other `bdrv_*()` functions except for the newly introduced ones,
making it a somewhat new QEMU object in the sense that it would
probably be more natural to have the user do something like this when
they want to create a new NVME device:
qemu \
    -ns-manager format=vssim,size=1M,id=my_manager \
    -device nvme,ns-manager=my_manager

Note that using this `-ns-manager` is what makes this a new kind of QEMU
object, it's not a device nor a drive.

Another approach that I was thinking about is letting the user specify a
drive for namespace management storage, like this:
qemu \
    -drive format=raw,file=ns_storage.img,id=ns_storage \
    -drive format=vssim,size=1M,id=vssim \
    -device nvme,drive=vssim,ns-drive=ns_storage

This way all the namespace management would be delegated to the nvme
device rather than the drive, and `ns-drive` would only store data
structures written direcly by the nvme device.

Change-Id: I9eb816b3b3336770a46f8cf87a66aaedc81308a0